+++ apps/wifi/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import WifiDetail, EmailDetail
+from .models import WifiDetail
 
 
 class WifiAdmin(admin.ModelAdmin):
@@ -7,11 +7,4 @@ class WifiAdmin(admin.ModelAdmin):
     search_fields = ('old_laptop_mac_address', 'new_laptop_mac_address', 'user__username',)
 
 
-class EmailAdmin(admin.ModelAdmin):
-    list_display = ('__unicode__', 'email', 'name', 'hod_approved_email', 'created', 'modified')
-    search_fields = ('email', 'name', 'email_purpose')
-    list_filter = ('hod_approved_email',)
-
-
 admin.site.register(WifiDetail, WifiAdmin)
-admin.site.register(EmailDetail, EmailAdmin)
